+++ ruby-lsp-ree/lib/ruby_lsp/ruby_lsp_ree/parsing/parsed_method_node.rb
@@ -14,6 +14,8 @@ class RubyLsp::Ree::ParsedMethodNode
   end
 
   def param_names
+    return [] unless @method_node.parameters
+    
     @method_node.parameters.requireds.map(&:name) + 
     @method_node.parameters.keywords.map(&:name) + 
     [@method_node.parameters.rest&.name] +